Added more docstrings, added fixes to bugs caused by refactor, and created a constants folder in src/Util
