--- src/Model/Tag/TagParams.php
@@ -1,7 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace ShipEngine\Model\Tag;
-
-final class TagParams
-{
-}
