--- src/Model/Tag/TagResult.php
@@ -1,14 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace ShipEngine\Model\Tag;
-
-use ShipEngine\Util;
-
-final class TagResult
-{
-    use Util\Getters;
-
-    public function __construct()
-    {
-    }
-}
