+++ src/Util/Constants/EventType.php
@@ -1,0 +1,21 @@
+<?php declare(strict_types=1);
+
+namespace ShipEngine\Util\Constants;
+
+/**
+ * Class EventType, this class provides constants used for events.
+ *
+ * @package ShipEngine\Util\Constants
+ */
+final class EventType
+{
+    /**
+     * @const RESPONSE_RECEIVED
+     */
+    const RESPONSE_RECEIVED = 'response_received';
+
+    /**
+     * @const REQUEST_SENT
+     */
+    const REQUEST_SENT = 'request_sent';
+}
