+++ src/Util/Constants/RPCMethods.php
@@ -1,0 +1,26 @@
+<?php declare(strict_types=1);
+
+namespace ShipEngine\Util\Constants;
+
+/**
+ * Class RPCMethods, this class provides constants use for RPC methods.
+ *
+ * @package ShipEngine\Util
+ */
+final class RPCMethods
+{
+    /**
+     * @const ADDRESS_VALIDATE
+     */
+    const ADDRESS_VALIDATE = 'address/validate';
+
+    /**
+     * @const PACKAGE_TRACK
+     */
+    const PACKAGE_TRACK = 'package/track';
+
+    /**
+     * @const CREATE_TAG
+     */
+    const CREATE_TAG = 'create/tag';
+}
