--- src/Util/RPCMethods.php
@@ -1,10 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace ShipEngine\Util;
-
-final class RPCMethods
-{
-    const ADDRESS_VALIDATE = 'address/validate';
-    const PACKAGE_TRACK = 'package/track';
-    const CREATE_TAG = 'create/tag';
-}
